continues

+++ initialize_stack.c
@@ -6,7 +6,7 @@
 /*   By: mpatrao <mpatrao@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/10 16:20:14 by mpatrao           #+#    #+#             */
-/*   Updated: 2023/01/12 14:55:40 by mpatrao          ###   ########.fr       */
+/*   Updated: 2023/01/12 15:22:03 by mpatrao          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,9 +27,9 @@ t_stack	*fill_stack(int ac, char **av)
 		if (value > INT_MAX || value < INT_MIN)
 			ft_error();
 		if (i == 1)
-			stack_a = stack_new_node((int)value);
+			stack_a = stack_new_node(value);
 		else
-			stack_add_bottom(&stack_a, stack_new_node((int)value));
+			stack_add_bottom(&stack_a, stack_new_node(value));
 		i++;
 	}
 	return (stack_a);
